requirements.txt
pydantic>=2.0,<2.5
langchain>=0.1.0
langchain-community
langchain-core
langchain-experimental
sqlalchemy
pymysql
python-dotenv
openai
langchain-openai

